dsc: add get dsc pix_per_clk [1/1]
PD#SWPL-142209

Problem:
vdin not get dsc pix_per_clk

Solution:
add get dsc pix_per_clk

Verify:
t3x

+++ drivers/media/vin/tvin/dsc_dec/dsc_dec_drv.h
@@ -62,7 +62,7 @@ struct aml_dsc_dec_drv_s {
 	unsigned int slice_num_m1;//pic_width/slice_width;
 	unsigned int dsc_dec_en;
 	unsigned int dsc_dec_frm_latch_en;// need to check ucode
-	unsigned int pix_per_clk;//input 0:1pix 1:2pix 2:4pix
+	u8 pix_per_clk;//input 0:1pix 1:2pix 2:4pix
 	bool c3_clk_en;
 	bool c2_clk_en;
 	bool c1_clk_en;
